chore : 오목 데이터베이스 패키지 수정
