+++ app/src/main/java/woowacourse/omok/data/db/omok/OmokDbHelper.kt
@@ -1,15 +1,18 @@
-package woowacourse.omok.data.db
+package woowacourse.omok.data.db.omok
 
 import android.content.Context
 import android.database.sqlite.SQLiteDatabase
 import android.database.sqlite.SQLiteOpenHelper
-import woowacourse.omok.data.db.OmokSchema.SQL_CREATE_ENTRIES
-import woowacourse.omok.data.db.OmokSchema.SQL_DELETE_ENTRIES
+import woowacourse.omok.data.db.omok.OmokSchema.SQL_CREATE_OMOK_TABLE
+import woowacourse.omok.data.db.omok.OmokSchema.SQL_DELETE_OMOK_TABLE
+import woowacourse.omok.data.db.room.RoomSchema.SQL_CREATE_ROOMS_TABLE
+import woowacourse.omok.data.db.room.RoomSchema.SQL_DELETE_ROOMS_TABLE
 
 class OmokDbHelper(context: Context) :
     SQLiteOpenHelper(context, DATABASE_NAME, null, DATABASE_VERSION) {
     override fun onCreate(db: SQLiteDatabase?) {
-        db?.execSQL(SQL_CREATE_ENTRIES)
+        db?.execSQL(SQL_CREATE_OMOK_TABLE)
+        db?.execSQL(SQL_CREATE_ROOMS_TABLE)
     }
 
     override fun onUpgrade(
@@ -17,7 +20,8 @@ class OmokDbHelper(context: Context) :
         oldVersion: Int,
         newVersion: Int,
     ) {
-        db?.execSQL(SQL_DELETE_ENTRIES)
+        db?.execSQL(SQL_DELETE_OMOK_TABLE)
+        db?.execSQL(SQL_DELETE_ROOMS_TABLE)
         onCreate(db)
     }
